pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = [{ name = "Nikolay Korkushko, email = <kakashonochka12@gmail.com>" }]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.1"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.11"
djangorestframework-simplejwt = "^5.5.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "black (>=25.9.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
test = [
    "coverage (>=7.11.0,<8.0.0)"
]
